+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## UNRELEASED
 ### Misc
 - make pre/post deprecation warnings more descriptive
+- remove pre/post from project configuration template
 - remove support for Ruby 2.2
 - bundler version constraint now supports bundler >= 2 (required by TravisCI)